backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.6
Flask==3.0.3
Flask-Cors==4.0.1
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-text-splitters==0.0.1
langsmith==0.1.56
MarkupSafe==2.1.5
marshmallow==3.21.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.3
packaging==23.2
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
SQLAlchemy==2.0.30
tenacity==8.3.0
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
yarl==1.9.4